Fix all warnings

+++ hopper/common.h
@@ -25,7 +25,7 @@ void drawLogo(void);
 void initTitle(void);
 bool updateTitle(void);
 void drawTitle(void);
-uint8_t setLastScore(int score, uint32_t time);
+uint8_t setLastScore(uint16_t score, uint32_t time);
 
 void initGame(void);
 bool updateGame(void);
